add: 練習11-2のプログラムを追加した
